+++ gelp/package.json
@@ -18,6 +18,7 @@
     "lucide-react": "^0.577.0",
     "mongoose": "^9.1.5",
     "next": "^16.1.6",
+    "postcss": "^8.5.10",
     "react": "19.2.3",
     "react-dom": "19.2.3",
     "sharp": "^0.34.5",
@@ -36,10 +37,8 @@
     "@types/react-dom": "^19",
     "jest": "^30.3.0",
     "jest-environment-jsdom": "^30.3.0",
-    "jest-environment-node": "^30.3.0",
     "mongodb-memory-server": "^11.0.1",
     "tailwindcss": "^4",
-    "ts-node": "^10.9.2",
     "tsx": "^4.21.0",
     "typescript": "^5"
   }
