+++ gelp/src/app/(auth_optional)/developer/page.tsx
@@ -1,0 +1,7 @@
+import Redirect from "@/components/util/Redirect";
+
+const Page = () => {
+  return <Redirect to={'/'}/>;
+}
+
+export default Page;
