+++ gelp/src/app/(auth_optional)/layout.tsx
@@ -20,8 +20,6 @@ const AuthOptionalLayout: FC<PropsWithChildren> = async ({children}) => {
     email: user.email,
     emailVerified: user.emailVerified,
     avatar: user.avatar,
-    createdAt: user.createdAt,
-    updatedAt: user.updatedAt,
     isAdministrator: user.isAdministrator
   } : null}>
     <Navbar />
