--- gelp/src/app/(authenticated)/authtest/page.tsx
@@ -1,11 +1,0 @@
-import UserDetails from "@/components/user/UserDetails";
-
-const Page = () => {
-  return <div>
-    <h1>Authenticated Successfully</h1>
-    <p>User details:</p>
-    <UserDetails/>
-  </div>
-}
-
-export default Page;
